pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperliquidmaster"
version = "0.1.0"
description = "Enhanced Hyperliquid Trading Bot"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "hyperliquid-python-sdk>=0.15.0",
    "pandas>=2.2.0",
    "numpy>=2.2.0",
    "requests>=2.32.0",
    "web3>=7.12.0",
    "eth-account>=0.13.0",
    "aiohttp>=3.7.4",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.6.0",
    "ta>=0.11.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "schedule>=1.1.0",
    "colorlog>=6.6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
ml = [
    "openai>=1.80.0",
    "nltk>=3.9.0",
    "textblob>=0.19.0",
]
dev = [
    "black",
    "ruff",
    "mypy",
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
hyperliquid = "hyperliquidmaster.scripts.master_bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
